// src/courier/pos.type.ts
export type PosType =
    ''
    | 'poslaju'
    | 'gdex'
    | 'jnt'
    | 'skynet'
    | 'citylink'
    | 'lex'
    | 'ninjavan'
    | 'dhl'
    | 'dhleco'
    | 'abx'
    | 'redpack'
    | 'cj'
    | 'fedex'
    | 'lwe'
    | 'lineclear'
    | 'airpak'
    | 'mypoz'
    | 'shopee'
    | 'taqbin'
    | 'ups'
    | 'zepto'
    | 'tnt'
    | 'pgeon'
    | 'aramex'
    | 'brazilianpost'
    | 'usps'
    | 'japan_ems'
    | 'posindonesia'
    | 'jne'
    | 'tiki'
    | 'sicepat'
    | 'wahana'
    | 'lion'
    | 'singpost'
    | 'nationwide'
    | 'aupost';